lxd: Move tests to `cloudinit` package
